+++ packages/formspree-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.8.1
+
+### Patch Changes
+
+- Catch Formspree errors using legacy format
+- Better error handling
+
 ## 2.8.0
 
 ### Minor Changes